CRYPTO_UTILS* constants moved to radix-engine-interface
This is to make them visible for scrypto
