+++ radix-engine-interface/src/blueprints/mod.rs
@@ -1,6 +1,7 @@
 pub mod access_controller;
 pub mod account;
 pub mod consensus_manager;
+pub mod crypto_utils;
 pub mod identity;
 pub mod package;
 pub mod pool;
